+++ src/tests/test_boptimizer.cpp
@@ -24,11 +24,11 @@ struct Params {
     };
 
     struct kf_maternfivehalfs {
-        BO_PARAM(float, sigma, 1);
-        BO_PARAM(float, l, 0.4);
+        BO_PARAM(double, sigma, 1);
+        BO_PARAM(double, l, 0.4);
     };
     struct ucb {
-        BO_PARAM(float, alpha, 0.125);
+        BO_PARAM(double, alpha, 0.125);
     };
     struct init {
         BO_PARAM(int, nb_samples, 5);
